refactor(store): further split mod.rs and tests.rs into focused modules
The remaining catch-all file and test suite has been split, separating
out different concerns into their own files.
